createCohorts {Reproducibility}	R Documentation

Create the exposure and outcome cohorts

Description

Create the exposure and outcome cohorts

Usage

createCohorts(
  connectionDetails,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable = "cohort",
  oracleTempSchema,
  outputFolder
)

Arguments

connectionDetails
An object of type connectionDetails as created using the createConnectionDetails function in the DatabaseConnector package.

cdmDatabaseSchema
Schema name where your patient-level data in OMOP CDM format resides. Note that for SQL Server, this should include both the database and schema name, for example 'cdm_data.dbo'.

cohortDatabaseSchema
Schema name where intermediate data can be stored. You will need to have write privileges in this schema. Note that for SQL Server, this should include both the database and schema name, for example 'cdm_data.dbo'.

cohortTable
The name of the table that will be created in the work database schema. This table will hold the exposure and outcome cohorts used in this study.

oracleTempSchema
Should be used in Oracle to specify a schema where the user has write privileges for storing temporary tables.

outputFolder
Name of local folder to place results; make sure to use forward slashes (/)

Details

This function will create the exposure and outcome cohorts following the definitions included in this package.

[Package Reproducibility version 0.0.1 Index]