exportResults {Reproducibility}	R Documentation

Export all results to tables

Description

Outputs all results to a folder called 'export', and zips them.

Usage

exportResults(
  outputFolder,
  databaseId,
  databaseName,
  databaseDescription,
  minCellCount = 5,
  maxCores
)

Arguments

outputFolder
Name of local folder to place results; make sure to use forward slashes (/). Do not use a folder on a network drive since this greatly impacts performance.

databaseId
A short string for identifying the database (e.g. 'Synpuf').

databaseName
The full name of the database.

databaseDescription
A short description (several sentences) of the database.

minCellCount
The minimum cell count for fields contains person counts or fractions.

maxCores
How many parallel cores should be used? If more cores are made available this can speed up the analyses.

[Package Reproducibility version 0.0.1 Index]