launchEvidenceExplorer {Reproducibility}	R Documentation

Launch the SqlRender Developer Shiny app

Description

Launch the SqlRender Developer Shiny app

Usage

launchEvidenceExplorer(dataFolder, blind = TRUE, launch.browser = TRUE)

Arguments

dataFolder
A folder where the data files for the Evidence Explorer app will be stored. Use the prepareForEvidenceExplorer to populate this folder.

blind
Should the user be blinded to the main results?

launch.browser
Should the app be launched in your default browser, or in a Shiny window. Note: copying to clipboard will not work in a Shiny window.

Details

Launches a Shiny app that allows the user to explore the evidence

[Package Reproducibility version 0.0.1 Index]