prepareForEvidenceExplorer {Reproducibility}	R Documentation

Prepare results for the Evidence Explorer Shiny app.

Description

Prepare results for the Evidence Explorer Shiny app.

Usage

prepareForEvidenceExplorer(resultsZipFile, dataFolder)

Arguments

resultsZipFile
Path to a zip file containing results from a study executed by this package.

dataFolder
A folder where the data files for the Evidence Explorer app will be stored.

Examples

## Not run: 
# Add results from three databases to the Shiny app data folder:
prepareForEvidenceExplorer("ResultsMDCD.zip", "/shinyData")
prepareForEvidenceExplorer("ResultsMDCR.zip", "/shinyData")
prepareForEvidenceExplorer("ResultsCCAE.zip", "/shinyData")

# Launch the Shiny app:
launchEvidenceExplorer("/shinyData")

## End(Not run)

[Package Reproducibility version 0.0.1 Index]