runCohortMethod {Reproducibility}	R Documentation

Run CohortMethod package

Description

Run CohortMethod package

Usage

runCohortMethod(
  connectionDetails,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable,
  oracleTempSchema,
  outputFolder,
  maxCores
)

Arguments

connectionDetails
An object of type connectionDetails as created using the createConnectionDetails function in the DatabaseConnector package.

cdmDatabaseSchema
Schema name where your patient-level data in OMOP CDM format resides. Note that for SQL Server, this should include both the database and schema name, for example 'cdm_data.dbo'.

cohortDatabaseSchema
Schema name where intermediate data can be stored. You will need to have write priviliges in this schema. Note that for SQL Server, this should include both the database and schema name, for example 'cdm_data.dbo'.

cohortTable
The name of the table that will be created in the work database schema. This table will hold the exposure and outcome cohorts used in this study.

oracleTempSchema
Should be used in Oracle to specify a schema where the user has write priviliges for storing temporary tables.

outputFolder
Name of local folder where the results were generated; make sure to use forward slashes (/). Do not use a folder on a network drive since this greatly impacts performance.

maxCores
How many parallel cores should be used? If more cores are made available this can speed up the analyses.

Details

Run the CohortMethod package, which implements the comparative cohort design.

[Package Reproducibility version 0.0.1 Index]