uploadResults {Reproducibility}	R Documentation

Upload results to OHDSI server

Description

Upload results to OHDSI server

Usage

uploadResults(outputFolder, privateKeyFileName, userName)

Arguments

outputFolder
Name of local folder where the results were generated; make sure to use forward slashes (/). Do not use a folder on a network drive since this greatly impacts performance.

privateKeyFileName
A character string denoting the path to the RSA private key provided by the study coordinator.

userName
A character string containing the user name provided by the study coordinator.

Details

This function uploads the 'Results_<databaseId>.zip' to the OHDSI SFTP server. Before sending, you can inspect the zip file, wich contains (zipped) CSV files. You can send the zip file from a different computer than the one on which is was created.

[Package Reproducibility version 0.0.1 Index]